add sessions-service intergration
