+++ Dockerfile.sessions-service
@@ -1,0 +1,28 @@
+### 1. Cloning stage
+FROM alpine/git:latest AS clone
+WORKDIR /workspace/sessions-service
+RUN git clone https://github.com/pagopa/pagopa-ecommerce-sessions-service
+
+### 2. Build stage
+FROM maven:3.8.3-openjdk-17 AS build
+WORKDIR /workspace/sessions-service/
+
+ARG REPO=/workspace/sessions-service/pagopa-ecommerce-sessions-service
+
+COPY --from=clone ${REPO}/ .
+RUN mvn -f ./pom.xml clean package
+
+### 3. Run stage
+FROM openjdk:17-jdk-alpine
+
+RUN addgroup -S spring && adduser -S spring -G spring
+USER spring:spring
+
+ARG EXTRACTED=/workspace/sessions-service/target/
+
+COPY --from=build --chown=spring ${EXTRACTED} .
+
+ARG JAR_FILE=*.jar
+ENV APP_NAME=$JAR_FILE
+
+CMD java -jar $APP_NAME
