+++ docker-compose.yml
@@ -4,12 +4,41 @@ services:
   pagopa-ecommerce-transactions:
     build:
       context: .
-      dockerfile: Dockerfile
     ports:
       - "8080:8080"
     networks:
       - pagopa-ecommerce
     env_file: .env
+
+  pagopa-ecommerce-sessions-service:
+    container_name: pagopa-ecommerce-sessions-service
+    build:
+      context: .
+      dockerfile: Dockerfile.sessions-service
+    ports:
+        - 8082:8080
+    networks:
+      - pagopa-ecommerce
+    
+  pagopa-node-mock:
+    container_name: pagopa-node-mock
+    build:
+      context: .
+      dockerfile: Dockerfile.nodo-mock
+    environment:
+      PORT: 3002
+      PAGOPA_PROXY_PORT: 3000
+      PAGOPA_PROXY_HOST: http://pagopaproxy
+      PAGOPA_NODO_HOST: http://pagopa-node-mock
+      PAGOPA_ID_CANALE: 97735020584_02
+      PAGOPA_ID_CANALE_PAGAMENTO: 97735020584_xx
+      PAGOPA_ID_INT_PSP: 97735020584
+      PAGOPA_ID_PSP: PAGOPA_01
+      PAGOPA_NODO_PASSWORD: password
+    ports:
+        - 3002:3002
+    networks:
+      - pagopa-ecommerce
 
   redis:
     image: redis
@@ -51,26 +80,9 @@ services:
       mongo:
         condition: service_healthy
 
-  pagopa-node-mock:
-    container_name: pagopa-node-mock
-    build:
-      context: .
-      dockerfile: Dockerfile.nodo-mock
-    environment:
-      PORT: 3002
-      PAGOPA_PROXY_PORT: 3000
-      PAGOPA_PROXY_HOST: http://pagopaproxy
-      PAGOPA_NODO_HOST: http://pagopa-node-mock
-      PAGOPA_ID_CANALE: 97735020584_02
-      PAGOPA_ID_CANALE_PAGAMENTO: 97735020584_xx
-      PAGOPA_ID_INT_PSP: 97735020584
-      PAGOPA_ID_PSP: PAGOPA_01
-      PAGOPA_NODO_PASSWORD: password
-    ports:
-        - 3002:3002
-    networks:
-      - pagopa-ecommerce    
-
 networks:
   pagopa-ecommerce:
     name: pagopa-ecommerce
+    driver: bridge
+    driver_opts:
+      com.docker.network.driver.mtu: 1450
