+++ src/main/java/it/pagopa/transactions/client/EcommerceSessionsClient.java
@@ -1,27 +1,32 @@
 package it.pagopa.transactions.client;
 
-import java.net.ConnectException;
-import java.time.Duration;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Component;
+import org.springframework.web.reactive.function.client.WebClient;
+import org.springframework.web.server.ResponseStatusException;
 
-import io.netty.handler.timeout.TimeoutException;
-import it.pagopa.ecommerce.sessions.v1.api.DefaultApi;
 import it.pagopa.ecommerce.sessions.v1.dto.SessionDataDto;
 import it.pagopa.ecommerce.sessions.v1.dto.SessionTokenDto;
 import reactor.core.publisher.Mono;
-import reactor.util.retry.Retry;
 
 @Component
 public class EcommerceSessionsClient {
 
     @Autowired
-    private DefaultApi ecommerceSessionsWebClient;
+    @Qualifier("ecommerceSessionsWebClient")
+    private WebClient ecommerceSessionsWebClient;
 
     public Mono<SessionTokenDto> createSessionToken(SessionDataDto request) {
 
-        return ecommerceSessionsWebClient.postToken(request).retryWhen(Retry.backoff(2, Duration.ofMillis(25))
-                .filter(throwable -> throwable instanceof TimeoutException || throwable instanceof ConnectException));
+        return ecommerceSessionsWebClient.post().body(Mono.just(request), SessionDataDto.class)
+                .retrieve()
+                .onStatus(HttpStatus::isError,
+                        clientResponse -> clientResponse.bodyToMono(String.class)
+                                .flatMap(errorResponseBody -> Mono.error(
+                                        new ResponseStatusException(clientResponse.statusCode(), errorResponseBody))))
+                .bodyToMono(SessionTokenDto.class);
     }
 }
