+++ src/main/java/it/pagopa/transactions/configuration/WebClientsConfig.java
@@ -12,7 +12,6 @@ import org.springframework.web.reactive.function.client.WebClient;
 import io.netty.channel.ChannelOption;
 import io.netty.handler.timeout.ReadTimeoutHandler;
 import it.pagopa.ecommerce.sessions.v1.ApiClient;
-import it.pagopa.ecommerce.sessions.v1.api.DefaultApi;
 import it.pagopa.nodeforpsp.ObjectFactory;
 import it.pagopa.transactions.utils.soap.Jaxb2SoapEncoder;
 import reactor.netty.http.client.HttpClient;
@@ -40,8 +39,30 @@ public class WebClientsConfig {
                 .build();
     }
 
+    // @Bean(name = "ecommerceSessionsWebClient")
+    // public DefaultApi
+    // ecommerceSessionsWebClient(@Value("${ecommerceSessions.uri}") String
+    // ecommerceSessionsUri,
+    // @Value("${ecommerceSessions.readTimeout}") int ecommerceSessionsReadTimeout,
+    // @Value("${ecommerceSessions.connectionTimeout}") int
+    // ecommerceSessionsConnectionTimeout) {
+
+    // HttpClient httpClient = HttpClient.create()
+    // .option(ChannelOption.CONNECT_TIMEOUT_MILLIS,
+    // ecommerceSessionsConnectionTimeout)
+    // .doOnConnected(connection -> connection
+    // .addHandlerLast(new ReadTimeoutHandler(ecommerceSessionsReadTimeout,
+    // TimeUnit.MILLISECONDS)));
+
+    // WebClient webClient = ApiClient.buildWebClientBuilder()
+    // .clientConnector(new
+    // ReactorClientHttpConnector(httpClient)).baseUrl(ecommerceSessionsUri).build();
+
+    // return new DefaultApi(new ApiClient(webClient));
+    // }
+
     @Bean(name = "ecommerceSessionsWebClient")
-    public DefaultApi ecommerceSessionsWebClient(@Value("${ecommerceSessions.uri}") String ecommerceSessionsUri,
+    public WebClient ecommerceSessionsWebClient(@Value("${ecommerceSessions.uri}") String ecommerceSessionsUri,
             @Value("${ecommerceSessions.readTimeout}") int ecommerceSessionsReadTimeout,
             @Value("${ecommerceSessions.connectionTimeout}") int ecommerceSessionsConnectionTimeout) {
 
@@ -50,10 +71,9 @@ public class WebClientsConfig {
                 .doOnConnected(connection -> connection
                         .addHandlerLast(new ReadTimeoutHandler(ecommerceSessionsReadTimeout, TimeUnit.MILLISECONDS)));
 
-        WebClient webClient = ApiClient.buildWebClientBuilder()
+        return ApiClient.buildWebClientBuilder()
                 .clientConnector(new ReactorClientHttpConnector(httpClient)).baseUrl(ecommerceSessionsUri).build();
 
-        return new DefaultApi(new ApiClient(webClient));
     }
 
     @Bean
