+++ src/main/resources/application.properties
@@ -5,6 +5,12 @@ logging.level.web=debug
 nodo.uri=${NODO_URI}
 nodo.readTimeout=${NODO_READ_TIMEOUT}
 nodo.connectionTimeout=${NODO_CONNECTION_TIMEOUT}
-ecommerceSessions.uri=${ECOMMERCE_SESSIONS_URI}
+ecommerceSessions.uri=http://pagopa-ecommerce-sessions-service:8082/sessions
 ecommerceSessions.readTimeout=${ECOMMERCE_SESSIONS_READ_TIMEOUT}
 ecommerceSessions.connectionTimeout=${ECOMMERCE_SESSIONS_CONNECTION_TIMEOUT}
+spring.data.mongodb.host=mongo
+spring.data.mongodb.port=27017
+spring.data.mongodb.authentication-database=admin
+spring.data.mongodb.username=admin
+spring.data.mongodb.password=password
+spring.data.mongodb.database=test
